KB-11711: resolved SonarQube code bug warnings. (#130)
* KB-11711: Added unit test cases to improve code coverage and resolved SonarQube code blocker warnings.

* Update CassandraConnectionMngrFactory.java

* KB-11711: Added unit test cases to improve code coverage and resolved SonarQube code blocker warnings.

* KB-11711: Added unit test cases to improve code coverage and resolved SonarQube code blocker warnings.

* KB-11711: Added unit test cases to improve code coverage and resolved SonarQube code blocker warnings.

* KB-11711: resolved SonarQube code blocker warnings.

* KB-11711: resolved SonarQube code blocker warnings.

* KB-11711: resolved SonarQube code bug warnings.

* KB-11711: resolved SonarQube code bug warnings(review comment resolved).

* KB-11711: resolved SonarQube code bug warnings(review comment resolved).

---------

+++ src/main/java/org/sunbird/workflow/utils/LRUCache.java
@@ -25,7 +25,7 @@ public class LRUCache<K, V> extends LinkedHashMap<K, V> {
     @PostConstruct
     public void postConstruct() {
         this.maxSize = conf.getEnrolStatusCountLocalCacheSize();
-        this.maxAgeInMillis = conf.getEnrolStatusCountLocalTimeToLive()*60;
+        this.maxAgeInMillis = (long) conf.getEnrolStatusCountLocalTimeToLive() * 60;
         this.entryTimeMap = new HashMap<>();
     }
     @Override
